refactor: (shield-447): update filename to kebab case
