+++ packages/mc-scripts/src/utils/credentials-storage.js
