+++ packages/mc-scripts/src/utils/validate-email.js
